Task Revamp complete: now task works with indexdb and new UI for date functionality
