--- to-do list copy/index.html
@@ -1,63 +1,0 @@
-<!DOCTYPE html>
-<html lang="en-us">
-    <head>
-        <title>To Do list</title>
-        <meta name="author" content="Dev" />
-        <link rel="stylesheet" href="./stylesheet.css" />
-        <meta name="viewport" content="width=device-width, initial-scale = 1.0" />
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-        <script src="https://kit.fontawesome.com/d9efd81c39.js" crossorigin="anonymous"></script>
-    </head>
-    <body class="body_background m-0 d-flex justify-content-center">
-        <div class="row w-100">
-            <div class="col-sm-2"></div>
-            <div class="bg-white col-sm-5 m-auto px-4 py-4 rounded task_content d-flex flex-column gap-3">
-                <div class="d-flex justify-content-between align-items-center">
-                    <div class="d-flex gap-2 align-items-center">
-                        <div class="fs-1" id="datepicker_date" >9</div>
-                        <div>
-                            <div id="datepicker_month">May</div>
-                            <div id="datepicker_year">2024</div>
-                        </div>
-                        <button id="DatePicker" class="datepicker-native btn btn-outline-secondary btn-sm rounded-circle">
-                            <i class="fa-regular fa-calendar"></i>
-                            <input type="date" class="datepicker-native-input" id="datepicker" onchange="handleDatePicker(event)" />
-                        </button>
-                    </div>
-                    <div class="fs-5" id="datepicker_day">Thursday</div>
-                </div>
-                <hr class="m-0 p-0"/>
-                <div id="task_list">
-                    <div id="box_1" class="box p-1" ondrop="drop(event)" ondragover="dragover(event)" ondragover="dragleave(event)">
-                        <div id="task_1" draggable="true" ondragstart="drag(event)" class="task fs-6 d-flex">
-                            <div class="flex-fill">Complete Project</div>
-                            <div class="d-flex gap-2">
-                                <div><i class="fa-solid fa-trash"></i></div>
-                                <div><i class="fa-regular fa-circle-check"></i></div>
-                            </div>
-                        </div>
-                    </div>
-                    <br/>
-                    <div id="box_2" class="box p-1" ondrop="drop(event)" ondragover="dragover(event)" ondragover="dragleave(event)">
-                        <div id="task_2" draggable="true" ondragstart="drag(event)" class="task fs-6 d-flex text-secondary">
-                            <div class="flex-fill">Complete Project</div>
-                            <div class="d-flex gap-2">
-                                <div><i class="fa-solid fa-trash"></i></div>
-                                <div><i class="fa-solid fa-circle-check text-success"></i></div>
-                            </div>
-                        </div>
-                    </div>
-                    <br/>
-                </div>
-                <div class="text-center justify-content-center d-flex gap-1">
-                    <div class="input-group">
-                        <input id="task-input" type="text" class="form-control border border-dark-subtle" >
-                    </div>  
-                    <button type="button" class="btn btn btn-outline-success" onclick="add_task()">Add</button>
-                </div>
-            </div>
-            <div class="col-sm-2"></div>
-        </div>
-        <script type="module" src="./script.js"></script>
-    </body>
-</html>
