--- to-do list copy/stylesheet.css
@@ -1,27 +1,0 @@
-.body_background{
-    background-color: #F0EFE9;
-    height: 100vh;
-}
-
-.datepicker-native {
-    position: relative;
-}
-.datepicker-native-input {
-    position: absolute;
-    overflow: hidden;
-    width: 100%;
-    height: 100%;
-    right: 0;
-    top: 0;
-    opacity: 0;
-}
-.datepicker-native-input::-webkit-calendar-picker-indicator {
-    position: absolute;
-    right: 0;
-    width: 100%;
-    height: 100%;
-    margin: 0;
-    padding: 0;
-    opacity: 0;
-    cursor: pointer;
-}
