+++ to-do list old/stylesheet.css
@@ -1,0 +1,32 @@
+.body_background{
+    background-image: linear-gradient(to bottom right, #1B3678, #290A32);
+    height: 100vh;
+}
+.add_button{
+    background-color: #FF653D;
+}
+.add_button:active{
+    background-color: #ed7c5f;
+}
+input[type="text"]:focus{
+    outline: none;
+}
+.hide {
+    display: none !important;
+}
+.dropover{
+    border: 1px solid blue;
+}
+.box{
+    height:5px;
+}
+.task_name{
+    word-wrap: break-word;
+    white-space: normal; 
+}
+.task_content{
+    max-height: 75vh;
+}
+.mouse_hover:hover{
+    cursor: pointer;
+}
