+++ to-do list/stylesheet.css
@@ -1,31 +1,29 @@
 .body_background{
-    background-image: linear-gradient(to bottom right, #1B3678, #290A32);
+    background-color: #F0EFE9;
     height: 100vh;
 }
-.add_button{
-    background-color: #FF653D;
+
+.datepicker-native {
+    position: relative;
 }
-.add_button:active{
-    background-color: #ed7c5f;
+.datepicker-native-input {
+    position: absolute;
+    overflow: hidden;
+    width: 100%;
+    height: 100%;
+    right: 0;
+    top: 0;
+    opacity: 0;
 }
-input[type="text"]:focus{
-    outline: none;
-}
-.hide {
-    display: none !important;
-}
-.dropover{
-    border: 1px solid blue;
-}
-.box{
-    height:5px;
-}
-.task_name{
-    word-wrap: break-word;
-    white-space: normal; 
-}
-.task_content{
-    max-height: 75vh;
+.datepicker-native-input::-webkit-calendar-picker-indicator {
+    position: absolute;
+    right: 0;
+    width: 100%;
+    height: 100%;
+    margin: 0;
+    padding: 0;
+    opacity: 0;
+    cursor: pointer;
 }
 .mouse_hover:hover{
     cursor: pointer;
